Avoid useSelector in Receive + Group selectors in reducers.
