--- src/redux/selectors/walletSelectors.ts
@@ -1,14 +1,0 @@
-import type { StateRestorerOpts } from 'ldk';
-
-import type { WalletState } from '../reducers/walletReducer';
-
-export function lastUsedIndexesSelector({
-  wallet,
-}: {
-  wallet: WalletState;
-}): StateRestorerOpts {
-  return {
-    lastUsedInternalIndex: wallet.lastUsedInternalIndex,
-    lastUsedExternalIndex: wallet.lastUsedExternalIndex,
-  };
-}
