+++ src/routes.ts
@@ -6,7 +6,6 @@ import Deposit from './pages/Deposit';
 import Faq from './pages/Faq';
 import Privacy from './pages/Privacy';
 import QRScanner from './pages/QRScanner';
-import Receive from './pages/Receive';
 import Settings from './pages/Settings';
 import ShowMnemonicSettings from './pages/ShowMnemonic/show-mnemonic-settings';
 import Terms from './pages/Terms';
@@ -16,6 +15,7 @@ import Exchange from './redux/containers/exchangeContainer';
 import TradeHistory from './redux/containers/exchangeHistoryContainer';
 import LiquidityProviders from './redux/containers/liquidityProvidersContainer';
 import Operations from './redux/containers/operationsContainer';
+import Receive from './redux/containers/receiveContainer';
 import ShowMnemonic from './redux/containers/showMnemonicContainer';
 import TradeSummary from './redux/containers/tradeSummaryContainer';
 import Wallet from './redux/containers/walletContainer';
